Kobo survey analysis from a data analysis plan

Kobo survey analysis from a data analysis plan

kobo_analysis_from_dap(
  design,
  dap,
  survey,
  choices,
  group = NULL,
  level = 0.95,
  choices_sep = "_"
)

Arguments

design

A srvyr::design object.

dap

A well formatted data analysis plan.#'

survey

The survey sheet from Kobo (with column "type" split). See split_survey().

choices

The choices sheet from Kobo.

group

A quoted vector of columns to group by. Default to NULL for no group.

level

(For vartype = "ci" only) A single number or vector of numbers indicating the confidence level

choices_sep

Select multiples choices separator.

Value

A survey-summarized-median data frame

Specifics

It takes the default for most variables, e.g. it tries to label, it calculates confidence intervals, and separators are the default.

Choices cannot be NULL.

The separator for choices can be changed.

Level can be changed.

See also